pyproject.toml
[project]
name = "vectorvfs"
version = "0.1.0"
description = "Your filesystem is a vector database."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "perception-models",
    "pillow>=11.0.0",
    "torch==2.5.1",
    "torchaudio==2.5.1",
    "torchcodec==0.1",
    "torchvision==0.20.1",
    "xformers>=0.0.29.post1",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "ruff>=0.11.7",
    "sphinx>=8.2.3",
    "sphinx-autobuild>=2024.10.3",
    "sphinx-press-theme>=0.9.1",
]

[tool.uv.sources]
perception-models = { git = "https://github.com/perone/perception_models.git", rev = "0389ab6823906760e3922a344dd2297c4e26ca94" }

[tool.mypy]
disable_error_code = ["import-untyped"]
